refactor(perf): extract pure helpers + wire test:perf:e2e-storm script
- Move span aggregation, fake-agent prompt building, MCP port discovery,
  graph-daemon node-bin resolution, vault stats, args parsing, exit-wait,
  summary printing, and userData seeding into a sibling helper module.
- The spec now reads as orchestration + Playwright fixtures only.
- Force-click into the project picker: cold-boot xvfb headful never
  resolves Playwright's `stable` actionability heuristic (the picker
  page is in continuous re-render); `force: true` bypasses it. If the
  click misses, the subsequent `.mcp.json` poll fails loudly.
- Add npm run test:perf:e2e-storm{,:local} mirroring perf:main-storm.
