+++ webapp/e2e-tests/electron/for_feature_development_not_LT_verification/graph/perf/electron-agent-storm-perf.spec.ts
@@ -1,89 +1,29 @@
 /**
  * E2E PERF TEST: real Electron app + real daemon + real agent storm.
  *
- * The diagonal cell that `agent-storm.ts` and the existing CDP perf specs do
- * NOT cover: a real Electron VoiceTree main bundle driven under an external
+ * Diagonal cell that `agent-storm.ts` and the existing CDP perf specs do NOT
+ * cover: a real Electron VoiceTree main bundle driven under an external
  * fake-agent storm. Daemon, electron-main, renderer, cytoscape, chokidar,
  * SQLite, and disk all run real; only the fake-agents are "mocked".
  *
  * Pipeline:
- *   1. Seed a temp `userData` + a fresh realistic vault via @vt/perf-fixtures.
- *   2. Launch `dist-electron/main/index.js` with --inspect=<port> and the
- *      seeded project. The app boots its own graph-db-server and writes
- *      <vault>/.mcp.json as it would for a real user.
- *   3. Read the MCP port from <vault>/.mcp.json (prod-shape discovery — no
- *      test-only IPC).
- *   4. Snapshot byte offsets of the three OTel NDJSON trace files
- *      (vt-graphd, vt-electron-daemon, vt-renderer).
- *   5. Start CDP main + renderer CPU profilers; record `Performance.getMetrics`
- *      before storm.
- *   6. Configure agent-runtime in this Playwright process (env-only, no
- *      daemon/MCP boot) and spawn N tmux-backed vt-fake-agents pointing at
- *      the discovered MCP port. Same env contract as `agent-storm.ts`.
- *   7. Await each agent's exit; record `Performance.getMetrics` after storm.
- *   8. Read NDJSON tails; aggregate spans; save .cpuprofile files; emit a
- *      JSON report at ~/.voicetree/reports/perf-agent-storm-e2e-<ts>.json
- *      with `args.variant: "e2e"` and the same top-level shape as the
- *      daemon-only harness's report (plus new `cdp.*` and `spans.vtRenderer`
- *      sections).
- *
- * Report shape (consumed by future dashboards; currently no consumer):
- *   args:                          parsed CLI flags + variant: "e2e"
- *   wallMs:                        storm wall time in ms
- *   agentCount/completedCount/...: same as daemon-only harness
- *   filesCreated:                  count of .md files in vault after storm
- *   vaultPath / appSupportPath:    absolute paths (kept if --keep-artifacts)
- *   spans.vtGraphd:                summary of vt-graphd.ndjson tail
- *   spans.vtElectronDaemon:        summary of vt-electron-daemon.ndjson tail
- *   spans.vtRenderer:              summary of vt-renderer.ndjson tail (zero
- *                                  counters when the file does not exist —
- *                                  the renderer tracer is a follow-up; see
- *                                  RENDERER SPAN SPEC below)
- *   cdp.mainMetrics:               Performance.getMetrics snapshots
- *                                  { before, after } for the electron main
- *                                  process inspector port
- *   cdp.rendererMetrics:           same for the renderer page CDP session
- *   cdp.profiles:                  absolute paths to main + renderer
- *                                  .cpuprofile files captured during the
- *                                  storm window
- *   agents[]:                      per-agent result objects
- *
- * RENDERER SPAN SPEC (deferred — the OpenSpec calls out renderer spans as a
- * follow-up change; this spec tolerates an absent vt-renderer.ndjson):
- *
- *   - `renderer.sse.event-received` — at
- *     webapp/src/shell/edge/UI-edge/text_to_tree_server_communication/sse-consumer.ts
- *     inside each `SSE_EVENT_TYPES` listener body. Attributes: `eventType`,
- *     `bytes`, `enqueuedAt`. Parent: extracted from `traceparent` if the
- *     daemon attaches it to SSE frame data (currently UNVERIFIED — the
- *     daemon's existing W3C propagator only injects on HTTP responses).
- *   - `renderer.delta.project` — wherever `projectDelta(delta)` is called in
- *     the renderer pipeline producing the `ProjectedGraph` for ui-apply.
- *     Attributes: `inputNodeCount`, `inputEdgeCount`, `outputNodeCount`,
- *     `outputEdgeCount`.
- *   - `renderer.delta.apply-ui` — wraps the body of `applyGraphDeltaToUI` in
- *     webapp/src/shell/edge/UI-edge/graph/actions/applyGraphDeltaToUI.ts.
- *     Attributes: `addedNodes`, `removedNodes`, `updatedNodes`, `addedEdges`.
- *   - `renderer.layout.invalidate` — wraps the call into
- *     `syncLargeGraphPerformanceMode` / cola layout invalidation triggered
- *     by a delta apply. Attributes: `dirtyNodeCount`, `mode` ('fullPack' |
- *     'localCola' | 'noop').
- *   - `renderer.react.render` — bracket around the next React commit after
- *     a delta apply. Attributes: `committedFiberCount` if cheaply available,
- *     else just timing.
- *
- *   Renderer tracer implementation (NOT in this change): a renderer-side
- *   OTel tracer using `@opentelemetry/sdk-trace-web` with an IPC-batched
- *   exporter that forwards spans to electron-main, which appends them to
- *   `~/.voicetree/traces/vt-renderer.ndjson` in the same NDJSON shape used
- *   by graph-db-server's existing tracer. The renderer tracer provider must
- *   register the same `CompositePropagator(W3CTraceContextPropagator,
- *   W3CBaggagePropagator)` so cross-process correlation continues to work.
+ *   1. Seed temp `userData` + a fresh realistic vault via @vt/perf-fixtures.
+ *   2. Launch `dist-electron/main/index.js` with --inspect=<port>.
+ *   3. Read MCP port from <vault>/.mcp.json (prod-shape discovery).
+ *   4. Snapshot byte offsets of the OTel NDJSON trace files.
+ *   5. Start CDP main + renderer CPU profilers.
+ *   6. Spawn N tmux-backed vt-fake-agents pointing at the discovered MCP port.
+ *   7. Await each agent's exit.
+ *   8. Read NDJSON tails, aggregate spans, save .cpuprofile files, emit
+ *      JSON report at ~/.voicetree/reports/perf-agent-storm-e2e-<ts>.json.
  *
  * Routing: invoked as `npm run test:perf:e2e-storm` (Onidel via run-remote.mjs)
- *   or `:local` for direct invocation. Linux hosts must have xvfb-run installed
- *   and the canonical xvfb wrapper in `packages/measures/src/_runners/
- *   run-with-xvfb-if-needed.ts` handles DISPLAY-less environments.
+ *   or `:local` for direct invocation. xvfb wrapped via
+ *   packages/measures/src/_runners/run-with-xvfb-if-needed.ts.
+ *
+ * Renderer span spec (deferred): see notes in
+ *   /Users/bobbobby/repos/voicetree-public/voicetree-26-5/stableperfprofilecollection.md
+ *   — the `vt-renderer` tracer is a follow-up; this spec tolerates its absence.
  */
 
 import { test as base, _electron as electron } from '@playwright/test'
@@ -94,15 +34,9 @@ import {
     existsSync,
     mkdirSync,
     mkdtempSync,
-    readFileSync,
-    readdirSync,
-    statSync,
     writeFileSync,
 } from 'node:fs'
 import * as os from 'node:os'
-import { execFileSync } from 'node:child_process'
-import { createRequire } from 'node:module'
-import { fileURLToPath } from 'node:url'
 
 import { generateVaultOnDisk, type VaultLayout } from '@vt/perf-fixtures'
 import { killOrphanVtGraphdDaemons } from '@vt/graph-db-client'
@@ -119,230 +53,26 @@ import {
     analyzeMainProcessProfile,
     printMainProcessMetrics,
 } from './perf-helpers/mainProcessProfile'
-
-const __filename = fileURLToPath(import.meta.url)
-const __dirname = path.dirname(__filename)
-const require = createRequire(import.meta.url)
-
-// ─── Args ────────────────────────────────────────────────────────────────
-
-interface E2EArgs {
-    readonly agents: number
-    readonly nodesPerAgent: number
-    readonly vaultSeedNodeCount: number
-    readonly perAgentTimeoutMs: number
-    readonly globalTimeoutMs: number
-    readonly keepArtifacts: boolean
-}
-
-function intEnv(key: string, fallback: number): number {
-    const raw = process.env[key]
-    if (raw === undefined || raw === '') return fallback
-    const n = Number.parseInt(raw, 10)
-    if (!Number.isInteger(n) || n < 0) throw new Error(`bad env ${key}=${raw}`)
-    return n
-}
-
-function boolEnv(key: string): boolean {
-    const raw = process.env[key]
-    return raw === '1' || raw === 'true'
-}
-
-function parseArgs(): E2EArgs {
-    return {
-        agents: intEnv('PERF_E2E_AGENTS', 8),
-        nodesPerAgent: intEnv('PERF_E2E_NODES_PER_AGENT', 30),
-        vaultSeedNodeCount: intEnv('PERF_E2E_VAULT_SEED_NODES', 300),
-        perAgentTimeoutMs: intEnv('PERF_E2E_PER_AGENT_TIMEOUT_MS', 120_000),
-        globalTimeoutMs: intEnv('PERF_E2E_GLOBAL_TIMEOUT_MS', 10 * 60_000),
-        keepArtifacts: boolEnv('PERF_E2E_KEEP_ARTIFACTS'),
-    }
-}
-
-// ─── Span / report helpers (mirrored from agent-storm.ts so the e2e and
-// daemon-only reports stay shape-compatible) ──────────────────────────────
-
-interface SpanRecord {
-    readonly traceId: string
-    readonly spanId: string
-    readonly name: string
-    readonly durationMs: number
-    readonly attributes?: Record<string, unknown>
-}
-
-interface SpanSummary {
-    readonly totalNew: number
-    readonly byName: Record<string, number>
-    readonly byOutcome: Record<string, number>
-    readonly durationsMs: Record<string, { p50: number; p95: number; p99: number; max: number }>
-}
-
-function ndjsonFileSize(filePath: string): number {
-    try { return statSync(filePath).size } catch { return 0 }
-}
-
-function readNdjsonTail(filePath: string, fromByteOffset: number): SpanRecord[] {
-    if (!existsSync(filePath)) return []
-    const buf = readFileSync(filePath)
-    if (buf.length <= fromByteOffset) return []
-    const tail = buf.subarray(fromByteOffset).toString('utf8')
-    const out: SpanRecord[] = []
-    for (const line of tail.split('\n')) {
-        const trimmed = line.trim()
-        if (trimmed.length === 0) continue
-        try {
-            out.push(JSON.parse(trimmed) as SpanRecord)
-        } catch {
-            // skip mid-write torn lines
-        }
-    }
-    return out
-}
-
-function quantile(sorted: readonly number[], q: number): number {
-    if (sorted.length === 0) return 0
-    if (sorted.length === 1) return sorted[0]
-    const pos = (sorted.length - 1) * q
-    const lo = Math.floor(pos)
-    const hi = Math.ceil(pos)
-    if (lo === hi) return sorted[lo]
-    return sorted[lo] * (1 - (pos - lo)) + sorted[hi] * (pos - lo)
-}
-
-function summarizeSpans(spans: readonly SpanRecord[]): SpanSummary {
-    const byName: Record<string, number> = {}
-    const byOutcome: Record<string, number> = {}
-    const durationsByName: Record<string, number[]> = {}
-    for (const span of spans) {
-        byName[span.name] = (byName[span.name] ?? 0) + 1
-        const outcome = span.attributes?.outcome
-        if (typeof outcome === 'string') {
-            const key = `${span.name}/${outcome}`
-            byOutcome[key] = (byOutcome[key] ?? 0) + 1
-        }
-        const list = durationsByName[span.name] ?? (durationsByName[span.name] = [])
-        list.push(span.durationMs)
-    }
-    const durationsMs: SpanSummary['durationsMs'] = {}
-    for (const [name, raw] of Object.entries(durationsByName)) {
-        const sorted = [...raw].sort((a, b) => a - b)
-        durationsMs[name] = {
-            p50: quantile(sorted, 0.5),
-            p95: quantile(sorted, 0.95),
-            p99: quantile(sorted, 0.99),
-            max: sorted[sorted.length - 1],
-        }
-    }
-    return { totalNew: spans.length, byName, byOutcome, durationsMs }
-}
-
-function countMarkdownFiles(dir: string): number {
-    let count = 0
-    const walk = (d: string): void => {
-        let entries: readonly string[]
-        try { entries = readdirSync(d) } catch { return }
-        for (const entry of entries) {
-            if (entry.startsWith('.')) continue
-            const full = path.join(d, entry)
-            let stat
-            try { stat = statSync(full) } catch { continue }
-            if (stat.isDirectory()) walk(full)
-            else if (stat.isFile() && entry.endsWith('.md')) count++
-        }
-    }
-    walk(dir)
-    return count
-}
-
-// ─── Fake-agent spawn helpers (re-implementing the daemon-harness contract,
-// adjusted for the Playwright-process caller) ─────────────────────────────
-
-function buildFakeAgentScript(nodesPerAgent: number): object {
-    const actions: object[] = []
-    for (let i = 0; i < nodesPerAgent; i++) {
-        actions.push({
-            type: 'create_node',
-            title: `Perf E2E Node ${i}`,
-            summary: `Synthetic node ${i} produced by e2e perf-agent-storm.`,
-            content: `Node body for index ${i}. Generated by the e2e perf harness.`,
-        })
-    }
-    actions.push({ type: 'exit', code: 0 })
-    return { actions }
-}
-
-function buildAgentPrompt(script: object): string {
-    return `### FAKE_AGENT_SCRIPT ###\n${JSON.stringify(script)}\n### END_FAKE_AGENT_SCRIPT ###`
-}
-
-function resolveRepoRoot(): string {
-    // webapp/e2e-tests/.../perf/electron-agent-storm-perf.spec.ts
-    // → ../../../../.. is the worktree root
-    return path.resolve(__dirname, '..', '..', '..', '..', '..', '..')
-}
-
-function resolveFakeAgentEntrypoint(): { dir: string; entry: string } {
-    const repoRoot = resolveRepoRoot()
-    const dir = path.join(repoRoot, 'tools', 'vt-fake-agent')
-    const entry = path.join(dir, 'src', 'index.ts')
-    if (!existsSync(entry)) throw new Error(`vt-fake-agent entrypoint not found at ${entry}`)
-    return { dir, entry }
-}
-
-function resolveTsxImportPath(): string {
-    return require.resolve('tsx')
-}
-
-// ─── MCP port discovery from .mcp.json (prod-shape contract) ─────────────
-
-async function readMcpPort(mcpJsonPath: string, timeoutMs: number): Promise<number> {
-    const deadline = Date.now() + timeoutMs
-    while (Date.now() < deadline) {
-        if (existsSync(mcpJsonPath)) {
-            try {
-                const raw = readFileSync(mcpJsonPath, 'utf8')
-                const parsed = JSON.parse(raw) as { mcpServers?: Record<string, { url?: string }> }
-                const url = parsed.mcpServers?.voicetree?.url
-                if (typeof url === 'string') {
-                    const m = url.match(/:(\d+)\/mcp$/)
-                    if (m) return Number.parseInt(m[1], 10)
-                }
-            } catch {
-                // file mid-write; try again
-            }
-        }
-        await new Promise(r => setTimeout(r, 250))
-    }
-    throw new Error(`timed out waiting for ${mcpJsonPath} with mcpServers.voicetree.url`)
-}
-
-// ─── graph-db-server native binding check (mirrors realistic-perf spec) ───
+import {
+    parseArgs,
+    type E2EArgs,
+    type AgentResult,
+    ndjsonFileSize,
+    readNdjsonTail,
+    summarizeSpans,
+    countMarkdownFiles,
+    buildFakeAgentScript,
+    buildAgentPrompt,
+    resolveFakeAgentEntrypoint,
+    resolveTsxImportPath,
+    readMcpPort,
+    resolveGraphDaemonNodeBin,
+    waitForExit,
+    printStormSummary,
+    seedUserData,
+} from './perf-helpers/agentStormE2eHelpers'
 
 const PROJECT_ROOT = path.resolve(process.cwd())
-
-function canLoadNativeGraphDbModules(nodeBin: string): boolean {
-    try {
-        execFileSync(nodeBin, ['-e', "const { DatabaseSync } = require('node:sqlite'); new DatabaseSync(':memory:').close()"], {
-            cwd: path.resolve(PROJECT_ROOT, '..'),
-            stdio: 'ignore',
-        })
-        return true
-    } catch {
-        return false
-    }
-}
-
-function resolveGraphDaemonNodeBin(): string {
-    const nvmNodeBin = path.join(os.homedir(), '.nvm', 'versions', 'node', 'v22.20.0', 'bin', 'node')
-    const candidates = [
-        process.env.VT_GRAPHD_NODE_BIN,
-        process.env.npm_node_execpath,
-        process.execPath,
-        existsSync(nvmNodeBin) ? nvmNodeBin : undefined,
-        'node',
-    ].filter((c): c is string => Boolean(c))
-    return candidates.find(canLoadNativeGraphDbModules) ?? process.execPath
-}
 
 // ─── Fixtures (one-shot, single test) ────────────────────────────────────
 
@@ -363,13 +93,9 @@ const test = base.extend<StormFixtures>({
     },
 
     vaultPath: async ({}, use) => {
-        // Vault sits inside a per-run temp project dir so the app's project
-        // picker has a reasonable name.
         const projectRoot = mkdtempSync(path.join(os.tmpdir(), 'vt-e2e-storm-project-'))
         const vault = path.join(projectRoot, 'perf-vault')
         await use(vault)
-        // Cleanup happens in `electronApp` teardown so the order is
-        // electron-close → vault-rm → orphan-reap.
     },
 
     appSupportPath: async ({}, use) => {
@@ -390,39 +116,7 @@ const test = base.extend<StormFixtures>({
     },
 
     electronApp: async ({ args, vaultPath, appSupportPath, vaultLayout: _vaultLayout }, use) => {
-        // Seed projects.json so the picker shows the vault project. We point
-        // `path` AND `voicetreeInitialized: true` at the vault so the app
-        // opens straight into graph view on click.
-        const projectsPath = path.join(appSupportPath, 'projects.json')
-        const projectName = path.basename(path.dirname(vaultPath))
-        await fs.writeFile(
-            projectsPath,
-            JSON.stringify([{
-                id: 'e2e-storm-perf-project',
-                path: path.dirname(vaultPath),
-                name: projectName,
-                type: 'folder',
-                lastOpened: Date.now(),
-                voicetreeInitialized: true,
-            }], null, 2),
-            'utf8',
-        )
-
-        const configPath = path.join(appSupportPath, 'voicetree-config.json')
-        await fs.writeFile(
-            configPath,
-            JSON.stringify({
-                lastDirectory: path.dirname(vaultPath),
-                vaultConfig: {
-                    [path.dirname(vaultPath)]: {
-                        writePath: vaultPath,
-                        readPaths: [],
-                    },
-                },
-            }, null, 2),
-            'utf8',
-        )
-
+        await seedUserData(fs, appSupportPath, vaultPath)
         const INSPECT_PORT = 9234
         const electronApp = await electron.launch({
             args: [
@@ -437,12 +131,11 @@ const test = base.extend<StormFixtures>({
                 MINIMIZE_TEST: process.env.MINIMIZE_TEST ?? '1',
                 VOICETREE_PERSIST_STATE: '1',
                 VOICETREE_DAEMON_LOAD_TIMEOUT_MS: '180000',
-                VT_GRAPHD_NODE_BIN: resolveGraphDaemonNodeBin(),
+                VT_GRAPHD_NODE_BIN: resolveGraphDaemonNodeBin(PROJECT_ROOT),
             },
             timeout: 30_000,
         })
 
-        // Surface [load-timing] from the bundled main.
         const mainStdout = electronApp.process().stdout
         if (mainStdout) {
             mainStdout.on('data', (chunk: Buffer) => {
@@ -472,7 +165,6 @@ const test = base.extend<StormFixtures>({
         }
         await electronApp.close()
 
-        // tear down agent-runtime tmux sessions before removing vault dir
         try { agentRuntime.getTerminalManager().cleanup() } catch { /* may be unconfigured if early failure */ }
 
         if (!args.keepArtifacts) {
@@ -501,15 +193,15 @@ const test = base.extend<StormFixtures>({
 
     mcpPort: async ({ vaultPath, appWindow }, use) => {
         // Click into the seeded project so the app opens the vault and writes
-        // .mcp.json. We target by `data-testid` rather than visible text — the
-        // text-match approach is brittle when the temp project name happens
-        // to span the layout (a 'visible / enabled / stable' check fails while
-        // the picker is still settling on a cold-boot Onidel disk).
+        // .mcp.json. `force: true` bypasses Playwright's `stable` actionability
+        // heuristic, which never resolves under cold-boot xvfb headful (the
+        // picker page is in a continuous re-render). If the click misses, the
+        // subsequent `.mcp.json` poll will fail loudly.
         await appWindow.waitForSelector('text=Voicetree', { timeout: 30_000 })
         const projectName = path.basename(path.dirname(vaultPath))
         const projectBtn = appWindow.locator('button[data-testid="saved-project-button"]').first()
         await projectBtn.waitFor({ state: 'visible', timeout: 30_000 })
-        await projectBtn.click({ timeout: 60_000 })
+        await projectBtn.click({ timeout: 60_000, force: true })
         console.log(`[E2E Storm] Clicked project '${projectName}' to enter graph view`)
 
         const mcpJsonPath = path.join(vaultPath, '.mcp.json')
@@ -519,50 +211,13 @@ const test = base.extend<StormFixtures>({
     },
 
     mainInspectPort: async ({}, use) => {
-        // Must match the value passed into --inspect= above. Hardcoded here
-        // mirrors the existing CDP perf spec; the only consumer of this value
-        // is the main-process CDP profiler, which polls /json/list on the port.
         await use(9234)
     },
 })
 
 // ─── The test ────────────────────────────────────────────────────────────
 
-interface AgentResult {
-    readonly terminalId: string
-    readonly spawnSuccess: boolean
-    readonly startedAtMs: number
-    readonly exitedAtMs: number | null
-    readonly exitCode: number | null
-    readonly stdoutSnippet: string
-    readonly errorMessage?: string
-}
-
-async function waitForExit(
-    terminalId: string,
-    exitedTerminals: Map<string, { code: number; atMs: number }>,
-    timeoutMs: number,
-): Promise<{ code: number; atMs: number } | null> {
-    const deadline = Date.now() + timeoutMs
-    while (Date.now() < deadline) {
-        const found = exitedTerminals.get(terminalId)
-        if (found) return found
-        const rec = agentRuntime.getTerminalRecords().find(r => r.terminalId === terminalId)
-        if (rec?.status === 'exited') {
-            const entry = { code: 0, atMs: Date.now() }
-            exitedTerminals.set(terminalId, entry)
-            return entry
-        }
-        await new Promise(r => setTimeout(r, 250))
-    }
-    return null
-}
-
 test.describe('E2E Electron + agent storm perf', () => {
-    // Fixture setup includes vite build (already done by `pretest`), Electron
-    // boot, project picker click, and .mcp.json polling — well over the config
-    // default of 90s on a cold Onidel disk. Override here so fixture setup is
-    // not strangled by the test-level timeout before we enter the body.
     test.describe.configure({ timeout: 15 * 60_000 })
 
     test('storm a real Electron-launched daemon with N fake-agents and report three-tracer + CDP timeline', async ({
@@ -575,12 +230,8 @@ test.describe('E2E Electron + agent storm perf', () => {
         mcpPort,
         mainInspectPort,
     }) => {
-        // ~3 min for boot + storm + report write; bumped if global timeout is.
         test.setTimeout(args.globalTimeoutMs + 90_000)
 
-        // ── Configure agent-runtime in THIS process for spawning only.
-        // No daemon, no MCP server, no graph bridge — those all live in the
-        // Electron app. We only need the tmux-backed terminal-manager.
         configureAgentRuntime({
             env: {
                 getAppSupportPath: () => appSupportPath,
@@ -590,7 +241,6 @@ test.describe('E2E Electron + agent storm perf', () => {
         await agentRuntime.ensureTmuxAvailable()
         await agentRuntime.ensureTmuxServer()
 
-        // ── Snapshot trace offsets BEFORE the storm window.
         const traceDir = path.join(os.homedir(), '.voicetree', 'traces')
         const graphdNdjson = path.join(traceDir, 'vt-graphd.ndjson')
         const electronNdjson = path.join(traceDir, 'vt-electron-daemon.ndjson')
@@ -599,7 +249,6 @@ test.describe('E2E Electron + agent storm perf', () => {
         const electronOffset = ndjsonFileSize(electronNdjson)
         const rendererOffset = ndjsonFileSize(rendererNdjson)
 
-        // ── CDP wiring: renderer profiler + main-process profiler + metric snapshots.
         const cdp = await appWindow.context().newCDPSession(appWindow)
         await cdp.send('Profiler.enable')
         await cdp.send('Performance.enable')
@@ -627,6 +276,8 @@ test.describe('E2E Electron + agent storm perf', () => {
 
         const script = buildFakeAgentScript(args.nodesPerAgent)
         const agentPrompt = buildAgentPrompt(script)
+        const getTerminalStatus = (id: string): string | undefined =>
+            agentRuntime.getTerminalRecords().find(r => r.terminalId === id)?.status
 
         // ── Storm window starts here.
         const wallStart = Date.now()
@@ -675,7 +326,7 @@ test.describe('E2E Electron + agent storm perf', () => {
                         errorMessage: spawn.error ?? 'spawn failed',
                     }
                 }
-                const exit = await waitForExit(terminalId, exitedTerminals, args.perAgentTimeoutMs)
+                const exit = await waitForExit(terminalId, exitedTerminals, getTerminalStatus, args.perAgentTimeoutMs)
                 const combined = (outputs.get(terminalId) ?? '')
                     + agentRuntime.getHeadlessAgentOutput(terminalId)
                 return {
@@ -717,7 +368,6 @@ test.describe('E2E Electron + agent storm perf', () => {
             metrics: Array<{ name: string; value: number }>
         }
 
-        // ── Stop renderer profiler; save .cpuprofile.
         const timestamp = new Date().toISOString().replace(/[:.]/g, '-')
         const reportsDir = path.join(os.homedir(), '.voicetree', 'reports')
         mkdirSync(reportsDir, { recursive: true })
@@ -731,7 +381,6 @@ test.describe('E2E Electron + agent storm perf', () => {
             printMainProcessMetrics(metrics)
         }
 
-        // ── Stop main-process profiler; save .cpuprofile.
         const mainProfilePath = await stopMainProcessProfileAndSave(
             reportsDir,
             `perf-agent-storm-e2e-main-${timestamp}.cpuprofile`,
@@ -741,7 +390,6 @@ test.describe('E2E Electron + agent storm perf', () => {
         console.log('\n[E2E Storm] MAIN PROCESS CPU PROFILE:')
         printMainProcessMetrics(mainMetrics)
 
-        // ── Tail trace files; summarize.
         const graphdSpans = readNdjsonTail(graphdNdjson, graphdOffset)
         const electronSpans = readNdjsonTail(electronNdjson, electronOffset)
         const rendererSpans = readNdjsonTail(rendererNdjson, rendererOffset)
@@ -771,9 +419,6 @@ test.describe('E2E Electron + agent storm perf', () => {
             },
             cdp: {
                 mainMetrics: {
-                    // Performance.getMetrics is only meaningful for the renderer
-                    // via this CDP session; main-process before/after surfaces are
-                    // captured implicitly via the .cpuprofile sampler.
                     note: 'See cdp.profiles.main for the sampling profile spanning the storm window',
                 },
                 rendererMetrics: {
@@ -791,35 +436,24 @@ test.describe('E2E Electron + agent storm perf', () => {
         const reportPath = path.join(reportsDir, `perf-agent-storm-e2e-${timestamp}.json`)
         writeFileSync(reportPath, JSON.stringify(report, null, 2), 'utf8')
 
-        // ── Console summary.
-        const sep = '='.repeat(70)
-        console.log(`\n${sep}`)
-        console.log('  E2E AGENT-STORM PERF — SUMMARY')
-        console.log(sep)
-        console.log(`  agents:        ${args.agents} requested, ${completedCount} ok, ${failedCount} failed, ${timedOutCount} timed out`)
-        console.log(`  nodes/agent:   ${args.nodesPerAgent}`)
-        console.log(`  wall time:     ${wallMs}ms`)
-        console.log(`  files written: ${filesCreated} (${(filesCreated / Math.max(1, wallMs / 1000)).toFixed(1)}/sec)`)
-        console.log(`  vt-graphd spans (new):           ${graphdSummary.totalNew}`)
-        for (const [name, n] of Object.entries(graphdSummary.byName).sort()) {
-            const d = graphdSummary.durationsMs[name]
-            console.log(`    ${name.padEnd(40)}  n=${String(n).padStart(5)}  p50=${d.p50.toFixed(1)}ms  p95=${d.p95.toFixed(1)}ms  max=${d.max.toFixed(1)}ms`)
-        }
-        console.log(`  vt-electron-daemon spans (new): ${electronSummary.totalNew}`)
-        for (const [name, n] of Object.entries(electronSummary.byName).sort()) {
-            const d = electronSummary.durationsMs[name]
-            console.log(`    ${name.padEnd(40)}  n=${String(n).padStart(5)}  p50=${d.p50.toFixed(1)}ms  p95=${d.p95.toFixed(1)}ms  max=${d.max.toFixed(1)}ms`)
-        }
-        console.log(`  vt-renderer spans (new):         ${rendererSummary.totalNew}`)
-        if (rendererSummary.totalNew === 0) {
-            console.log('    (renderer tracer not yet implemented — see RENDERER SPAN SPEC at top of file)')
-        }
-        const heapBefore = cdpRendererBefore.metrics.find(m => m.name === 'JSHeapUsedSize')?.value ?? 0
-        const heapAfter = cdpRendererAfter.metrics.find(m => m.name === 'JSHeapUsedSize')?.value ?? 0
-        console.log(`  renderer JS heap (after):        ${(heapAfter / 1024 / 1024).toFixed(1)} MB (delta ${((heapAfter - heapBefore) / 1024 / 1024).toFixed(1)} MB)`)
-        console.log(`  report:        ${reportPath}`)
-        console.log(`  cpuprofiles:   ${mainProfilePath}\n                 ${rendererProfilePath}`)
-        console.log(sep)
+        const heapBeforeBytes = cdpRendererBefore.metrics.find(m => m.name === 'JSHeapUsedSize')?.value ?? 0
+        const heapAfterBytes = cdpRendererAfter.metrics.find(m => m.name === 'JSHeapUsedSize')?.value ?? 0
+        printStormSummary({
+            args,
+            completedCount,
+            failedCount,
+            timedOutCount,
+            wallMs,
+            filesCreated,
+            graphdSummary,
+            electronSummary,
+            rendererSummary,
+            heapBeforeBytes,
+            heapAfterBytes,
+            reportPath,
+            mainProfilePath,
+            rendererProfilePath,
+        })
 
         // ── v1 assertions: structural only, no latency thresholds.
         if (agentResults.length !== args.agents) {
@@ -839,8 +473,6 @@ test.describe('E2E Electron + agent storm perf', () => {
                 + 'electron-main observability did not run, or the trace file is wrong',
             )
         }
-        // No assertion on rendererSummary.totalNew — that requires the renderer
-        // tracer follow-up.
     })
 })
